Trait pallet_quorum::pallet::Config

pub trait Config: Config + Config<AssetId = AssetId, Balance = Balance> {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type QuorumPalletId: Get<PalletId>;
    type ProposalsCap: Get<u32>;
    type BurnedCap: Get<u32>;
    type ProposalLifetime: Get<Self::BlockNumber>;
    type WeightInfo: WeightInfo;
    type Security: SecurityExt<Self::AccountId, Self::BlockNumber>;
    type StringLimit: Get<u32>;
    type VotesLimit: Get<u32>;
    type WatchListLimit: Get<u32>;
    type PubkeyLimitPerAsset: Get<u32>;
    type AssetRegistry: AssetRegistryExt;
    type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;

}

Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Events

type QuorumPalletId: Get<PalletId>

Pallet ID

type ProposalsCap: Get<u32>

Proposals capacity

type BurnedCap: Get<u32>

Burned queue capacity

type ProposalLifetime: Get<Self::BlockNumber>

Proposals lifetime

type WeightInfo: WeightInfo

Weights

type Security: SecurityExt<Self::AccountId, Self::BlockNumber>

Security traits

type StringLimit: Get<u32>

The maximum length of string (public keys etc..)

type VotesLimit: Get<u32>

The maximum number of votes per proposal

type WatchListLimit: Get<u32>

The maximum number of proposals per account watch list

type PubkeyLimitPerAsset: Get<u32>

The pubkey per asset (should always be more than current member size)

type AssetRegistry: AssetRegistryExt

Asset registry traits

type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Tidechain currency wrapper